Relocate SleuthSpanContextSupplier
